refactor(notebook): move read-only cell data to shell boundary
